Move _sub_symbols to _operations
